backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
openai>=1.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
pillow>=10.2.0
faster-whisper>=1.0.0
supabase>=2.24.0

